feat(st-clock): add barriers in clock driver
Add barriers in the clock drivers.
Even if peripheral memory is non-cacheable and the order of operations
sent on the bus is guaranteed, there is no guarantee regarding the order
of reception on the device. This implies that the order of accesses
arriving at the device memory is guaranteed only within the same
peripheral. Add a DMB before disabling a clock. And re-read the RCC
register after enabling a clock to be sure the transaction has reached
the clock peripheral.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: Ie37409173a3e3f4f4acf96a5ada8ae1b5a0c6784