feat(qemu-sbsa): Make RMM manifest DRAM information dynamic
The RMM manifest DRAM information is currently hard coded, something
that doesn't work when the base and size of the DRAM is modified at
runtime.  Such a case is when the RMM if located at the bottom of the
DRAM and the base address has to be shifted upward and the size reduced.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>