fix(nxp): move includes to where they are needed
The PSCI platform code relies on various headers being included from
common code and includes very little itself. As a result, changes in
common code cause the platform build to fail. Add some headers that are
necessary for the compilation of the file and move the core includes to
where they are used.

Change-Id: I7dc10e62cf71e5c70bcb2a874ffc7785a24189fd
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>